feat: Add Route to retrieve notas
- Add routes to scrap notas, faltas
- Add route to get notas
- Add validation to routes
